사용자 위치 즐겨찾기 업데이트 기능 구현 사용자 위치 설명 업데이트 기능 구현 사용자 상세 페이지 설계

+++ src/main/java/com/teamname/astroneer/star_info_web/security/CustomOAuth2User.java
@@ -65,6 +65,6 @@ public class CustomOAuth2User implements OAuth2User, UserDetails {
         return member.getEmail();
     }
 
-    public int getUserId() { return member.getId();
+    public long getUserId() { return member.getId();
     }
 }
